Moved resource node queries to `/api/http/resourceNodes.ts`
